Pull getSecret out into its own module in both functions that use it
